sprites from spritesheets
addet an function to the BlockType class that returns the sprite of that block. It does this by knowing how big one tile is and then getting the correct part of the tileset that tilemap uses aswell. Basically no longer is it neccesary that for each new block there needs to be an extra sprite, it just gets generated from the spritesheet which needs to be updates anyways for said new block. Half as much clicking. Also fixed the scuffed blockswap curser logic.